refactor: prover backend
