+++ prover-ray/backend/zkc-r5/zkc_r5_test.go
@@ -1,0 +1,187 @@
+package zkc_r5_test
+
+import (
+	"bytes"
+	"encoding/binary"
+	"math"
+	"testing"
+
+	zkc_r5 "github.com/LFDT-Lineth/lineth-monorepo/prover-ray/backend/zkc-r5"
+	minimal_elf "github.com/LFDT-Lineth/lineth-monorepo/prover-ray/internal/minimal-elf"
+	"github.com/stretchr/testify/assert"
+	"github.com/stretchr/testify/require"
+)
+
+func TestSszBlobs_LengthPrefixAtInOrigin(t *testing.T) {
+	ssz := []byte{0xAA, 0xBB, 0xCC}
+	got, err := zkc_r5.NewDataSection(zkc_r5.DefaultINOrigin, ssz)
+	require.NoError(t, err)
+	require.Len(t, got, 2)
+
+	// First memory blob: 8-byte LE length at inOrigin.
+	assert.Equal(t, zkc_r5.DefaultINOrigin, got[0].Offset, "first memory blob offset must be inOrigin")
+	require.Len(t, got[0].Data, 8, "length prefix must be exactly 8 bytes")
+	assert.Equal(t, uint64(3), binary.LittleEndian.Uint64(got[0].Data), "length prefix must encode payload length as LE uint64")
+}
+
+func TestSszBlobs_PayloadAtInOriginPlus8(t *testing.T) {
+	payload := []byte{0xAA, 0xBB, 0xCC}
+	got, err := zkc_r5.NewDataSection(zkc_r5.DefaultINOrigin, payload)
+	require.NoError(t, err)
+	require.Len(t, got, 2)
+
+	// Second memory blob: the payload bytes at inOrigin+8.
+	assert.Equal(t, zkc_r5.DefaultINOrigin+8, got[1].Offset, "payload memory blob offset must be inOrigin+8")
+	assert.Equal(t, payload, got[1].Data, "payload memory blob must contain the payload bytes verbatim")
+}
+
+func TestSszBlobs_EmptySSZ(t *testing.T) {
+	// Empty SSZ: only the 8-byte length memory blob, no payload memory blob.
+	got, err := zkc_r5.NewDataSection(zkc_r5.DefaultINOrigin, nil)
+	require.NoError(t, err)
+	require.Len(t, got, 1, "empty SSZ must produce exactly one memory blob (length prefix only)")
+	assert.Equal(t, zkc_r5.DefaultINOrigin, got[0].Offset, "length prefix memory blob offset must be inOrigin")
+	assert.Equal(t, uint64(0), binary.LittleEndian.Uint64(got[0].Data), "length prefix must be zero for empty SSZ")
+}
+
+func TestBuildZkcInputs_SSZOffsetOverflow(t *testing.T) {
+	elfBytes := minimal_elf.Make(minimal_elf.DefaultEntryPoint, minimal_elf.DefaultSectionAddr, minimal_elf.ValidSectionData)
+	_, err := zkc_r5.PrepareInput(elfBytes, []byte{0x01}, math.MaxUint64-4)
+	require.Error(t, err)
+	assert.Contains(t, err.Error(), "data input offset overflow")
+	assert.Contains(t, err.Error(), "in_origin=0xfffffffffffffffb")
+}
+
+func TestEncodeInputs_EntryPointAndCount(t *testing.T) {
+	const entryPoint = 0x11223344
+	inputs := zkc_r5.EncodeGuestAndMemoryForZkc(
+		zkc_r5.GuestProgramSections{
+			EntryPoint: entryPoint,
+		},
+		[]zkc_r5.ElfSection{
+			{Offset: 0x1000, Data: []byte{0x01}},
+			{Offset: 0x2000, Data: []byte{0x02}},
+		},
+	)
+
+	got := inputs["entry_point_and_blobs_count"]
+	require.Len(t, got, 16, "entry_point_and_blobs_count must be two BE uint64s")
+	assert.Equal(t, uint64(entryPoint), binary.BigEndian.Uint64(got[:8]), "first 8 bytes must be the BE entry point")
+	assert.Equal(t, uint64(2), binary.BigEndian.Uint64(got[8:]), "last 8 bytes must be the BE blob count")
+}
+
+func TestEncodeInputs_OffsetAndSizePairs(t *testing.T) {
+	memBlobs := []zkc_r5.ElfSection{
+		{Offset: 0x00800000, Data: make([]byte, 0x1234)},
+		{Offset: 0x2000, Data: []byte{0xBB}},
+	}
+	inputs := zkc_r5.EncodeGuestAndMemoryForZkc(zkc_r5.GuestProgramSections{}, memBlobs)
+
+	got := inputs["blobs_offset_and_size"]
+	require.Len(t, got, 32, "blobs_offset_and_size must be one 16-byte pair per blob")
+	assert.Equal(t, uint64(0x00800000), binary.BigEndian.Uint64(got[0:8]))
+	assert.Equal(t, uint64(0x1234), binary.BigEndian.Uint64(got[8:16]))
+	assert.Equal(t, uint64(0x2000), binary.BigEndian.Uint64(got[16:24]))
+	assert.Equal(t, uint64(1), binary.BigEndian.Uint64(got[24:32]))
+}
+
+func TestEncodeInputs_DataConcatenated(t *testing.T) {
+	memBlobs := []zkc_r5.ElfSection{
+		{Offset: 0x1000, Data: []byte{0xAA, 0xBB}},
+		{Offset: 0x2000, Data: []byte{0xCC}},
+	}
+	inputs := zkc_r5.EncodeGuestAndMemoryForZkc(zkc_r5.GuestProgramSections{}, memBlobs)
+	assert.Equal(t, []byte{0xAA, 0xBB, 0xCC}, inputs["blobs_data"], "blobs_data must be all blob bytes concatenated in order")
+}
+
+func TestElfBlobs_ExtractsSectionAtCorrectOffset(t *testing.T) {
+	parsedELF, err := zkc_r5.LoadGuestElf(bytes.NewReader(minimal_elf.MinimalElfProgram))
+	require.NoError(t, err)
+	require.Len(t, parsedELF.Sections, 1, "one loadable section must yield one memory blob")
+	assert.Equal(t, uint64(minimal_elf.DefaultSectionAddr), parsedELF.Sections[0].Offset, "memory blob offset must match the section's virtual address")
+	assert.Equal(t, minimal_elf.ValidSectionData, parsedELF.Sections[0].Data, "memory blob data must match the section bytes")
+}
+
+func TestElfBlobs_EntryPointPreserved(t *testing.T) {
+	parsedELF, err := zkc_r5.LoadGuestElf(bytes.NewReader(minimal_elf.MinimalElfProgram))
+	require.NoError(t, err)
+	assert.Equal(t, uint64(minimal_elf.DefaultEntryPoint), parsedELF.EntryPoint, "entry point must match ELF e_entry")
+}
+
+func TestElfBlobs_InvalidELFReturnsError(t *testing.T) {
+	_, err := zkc_r5.LoadGuestElf(bytes.NewReader([]byte("not an elf")))
+	assert.Error(t, err, "malformed ELF must return an error")
+}
+
+func TestBuildZkcInputs_ReturnsThreeKeys(t *testing.T) {
+	// These three key names are declared in RISCV-ZKC.bin's main.zkc. A name
+	// mismatch causes a silent no-op when the prover loads inputs.
+	inputs, err := zkc_r5.PrepareInput(minimal_elf.MinimalElfProgram, []byte{0x01, 0x02}, zkc_r5.DefaultINOrigin)
+	require.NoError(t, err)
+
+	assert.Contains(t, inputs, "entry_point_and_blobs_count")
+	assert.Contains(t, inputs, "blobs_offset_and_size")
+	assert.Contains(t, inputs, "blobs_data")
+}
+
+// TestBuildZkcInputs_NoLoadableSegments verifies that an ELF whose only
+// segment is not PT_LOAD is rejected with a clear error rather than producing
+// an empty blob set.
+func TestBuildZkcInputs_NoLoadableSegments(t *testing.T) {
+	elfBytes := minimal_elf.Make(minimal_elf.DefaultEntryPoint, minimal_elf.DefaultSectionAddr, minimal_elf.ValidSectionData)
+	// The program header starts at byte 64; its first field is p_type
+	// (uint32 LE). Rewrite PT_LOAD (1) to PT_NOTE (4).
+	binary.LittleEndian.PutUint32(elfBytes[64:68], 4)
+
+	_, err := zkc_r5.PrepareInput(elfBytes, []byte{0x01}, zkc_r5.DefaultINOrigin)
+	require.Error(t, err)
+	assert.Contains(t, err.Error(), "no loadable sections")
+}
+
+func TestBuildZkcInputs_LoadableSegmentFileSizeExceedsMemSize(t *testing.T) {
+	elfBytes := minimal_elf.Make(minimal_elf.DefaultEntryPoint, minimal_elf.DefaultSectionAddr, minimal_elf.ValidSectionData)
+	// Program header starts at byte 64. p_filesz is at +32, p_memsz at +40.
+	binary.LittleEndian.PutUint64(elfBytes[64+32:64+40], 2)
+	binary.LittleEndian.PutUint64(elfBytes[64+40:64+48], 1)
+
+	_, err := zkc_r5.PrepareInput(elfBytes, []byte{0x01}, zkc_r5.DefaultINOrigin)
+	require.Error(t, err)
+	assert.Contains(t, err.Error(), "file size larger than memory size")
+}
+
+func TestBuildZkcInputs_LoadableSegmentAddressOverflow(t *testing.T) {
+	elfBytes := minimal_elf.Make(minimal_elf.DefaultEntryPoint, minimal_elf.DefaultSectionAddr, minimal_elf.ValidSectionData)
+	// Program header starts at byte 64. p_vaddr is at +16, p_memsz at +40.
+	binary.LittleEndian.PutUint64(elfBytes[64+16:64+24], math.MaxUint64-1)
+	binary.LittleEndian.PutUint64(elfBytes[64+40:64+48], 4)
+
+	_, err := zkc_r5.PrepareInput(elfBytes, []byte{0x01}, zkc_r5.DefaultINOrigin)
+	require.Error(t, err)
+	assert.Contains(t, err.Error(), "segment address overflow")
+}
+
+func TestBuildZkcInputs_SectionAddressOverflow(t *testing.T) {
+	elfBytes := minimal_elf.Make(minimal_elf.DefaultEntryPoint, minimal_elf.DefaultSectionAddr, minimal_elf.ValidSectionData)
+	shOff := binary.LittleEndian.Uint64(elfBytes[40:48])
+	textSection := shOff + 64 // section 0 is NULL, section 1 is .text
+	// Section header field offsets: sh_addr at +16, sh_size at +32.
+	binary.LittleEndian.PutUint64(elfBytes[textSection+16:textSection+24], math.MaxUint64-1)
+	binary.LittleEndian.PutUint64(elfBytes[textSection+32:textSection+40], 4)
+
+	_, err := zkc_r5.PrepareInput(elfBytes, []byte{0x01}, zkc_r5.DefaultINOrigin)
+	require.Error(t, err)
+	assert.Contains(t, err.Error(), "section .text address overflow")
+}
+
+func TestBuildZkcInputs_SectionShortRead(t *testing.T) {
+	elfBytes := minimal_elf.Make(minimal_elf.DefaultEntryPoint, minimal_elf.DefaultSectionAddr, minimal_elf.ValidSectionData)
+	shOff := binary.LittleEndian.Uint64(elfBytes[40:48])
+	textSection := shOff + 64 // section 0 is NULL, section 1 is .text
+	// Section header field offsets: sh_offset at +24, sh_size at +32.
+	binary.LittleEndian.PutUint64(elfBytes[textSection+24:textSection+32], uint64(len(elfBytes)-2))
+	binary.LittleEndian.PutUint64(elfBytes[textSection+32:textSection+40], 4)
+
+	_, err := zkc_r5.PrepareInput(elfBytes, []byte{0x01}, zkc_r5.DefaultINOrigin)
+	require.Error(t, err)
+	assert.Contains(t, err.Error(), "short read for section .text")
+}
